Inject the entity manager at startup...
 ... Freeing AssetManager, EntityExecution, EntityQueryBuilder, and Interpreter
 from cycles
